requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.2
python-dotenv==1.0.1
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
google-generativeai
pgvector
